Clean structure, intercept posthog
