+++ extensions/pseudocontext/chunker/__init__.py
@@ -1,0 +1,3 @@
+from .instruct import InstructChunker
+from .naive import NaiveChunker
+from .paragraph import ParagraphChunker
