+++ extensions/pseudocontext/chunker/paragraph.py
@@ -1,0 +1,7 @@
+from ..base import Chunker
+
+
+class ParagraphChunker(Chunker):
+    """
+    Paragraph chunker creates chunks by spliting on newlines
+    """
