+++ extensions/pseudocontext/collector/__init__.py
@@ -1,0 +1,1 @@
+from .chroma import ChromaCollector
