+++ extensions/pseudocontext/collector/chroma.py
@@ -1,0 +1,27 @@
+import posthog
+
+print('Intercepting all calls to posthog :)')
+posthog.capture = lambda *args, **kwargs: None
+
+import chromadb
+from chromadb.config import Settings
+from ..base import Collecter, Embedder
+
+class ChromaCollector(Collecter):
+    def __init__(self, embedder: Embedder):
+        super().__init__()
+        self.chroma_client = chromadb.Client(Settings(anonymized_telemetry=False))
+        self.embedder = embedder
+        self.collection = self.chroma_client.create_collection(name="context", embedding_function=embedder.embed)
+        self.ids = []
+    
+    def add(self, texts: list[str]):
+        self.ids = [f"id{i}" for i in range(len(texts))]
+        self.collection.add(documents=texts, ids=self.ids)
+        
+    def get(self, search_strings: list[str], n_results: int) -> list[str]:
+        result = self.collection.query(query_texts=search_strings, n_results=n_results, include=['documents'])['documents'][0]
+        return result
+    
+    def clear(self):
+        self.collection.delete(ids=self.ids)
