+++ extensions/pseudocontext/embedder/__init__.py
@@ -1,0 +1,1 @@
+from .sentence_transformer import SentenceTransformerEmbedder
