+++ extensions/pseudocontext/embedder/sentence_transformer.py
@@ -1,0 +1,8 @@
+from sentence_transformers import SentenceTransformer
+from ..base import Embedder
+
+
+class SentenceTransformerEmbedder(Embedder):
+    def __init__(self) -> None:
+        self.model = SentenceTransformer("sentence-transformers/all-mpnet-base-v2")
+        self.embed = self.model.encode
