+++ extensions/pseudocontext/injector/__init__.py
@@ -1,0 +1,2 @@
+from .instruct import InstructInjector
+from .naive import NaiveInjector
