+++ extensions/pseudocontext/retriever/__init__.py
@@ -1,0 +1,2 @@
+from .cosine_similarity import CosineSimilarityRetriever
+from .instruct import InstructRetriever
